+++ polls/serializers.py
@@ -10,35 +10,38 @@ class OptionSerializer(ModelFileSerializer):
 
     class Meta:
         model = Option
-        fields = ('id', 'poll', 'label', 'media')
-        read_only_fields = ('id', 'poll')
+        fields = ('id', 'label', 'media')
+        read_only_fields = ('id',)
+
+    def validate(self, attrs):
+        user = self.context['request'].user
+        poll = self.context['poll']
+        if user != poll.creator:
+            raise serializers.ValidationError("You cannot add option to other user's poll")
+        return attrs
+
+    def create(self, validated_data):
+        poll = self.context['poll']
+        media = validated_data.pop('media')
+        option = Option.objects.create(**validated_data, media=media['name'], poll=poll)
+        return option
+
+    def update(self, instance, validated_data):
+        media = validated_data.pop('media', None)
+        if media:
+            validated_data['media'] = media['name']
+        return super().update(instance, validated_data)
 
 
-class PollSerializer(serializers.ModelSerializer):
-    options = OptionSerializer(many=True)
+class PollSerializer(ModelFileSerializer):
+    image = FileUploadSerializer(read_only=True)
 
     class Meta:
         model = Poll
-        fields = ('id', 'title', 'places_number', 'media_type', 'options')
-
-    def validate_places_number(self, value):
-        if value <= 0 or value % 2 != 0:
-            raise serializers.ValidationError('Places should be even number')
-        return value
-
-    def validate(self, data):
-        if len(data['options']) != data['places_number']:
-            raise serializers.ValidationError('Number of options should be equal to number of places')
-        return super().validate(data)
+        fields = ('id', 'title', 'media_type', 'description', 'image')
 
     @transaction.atomic
     def create(self, validated_data):
         user = self.context['request'].user
-        options = validated_data.pop('options')
         poll = Poll.objects.create(**validated_data, creator=user)
-
-        for option in options:
-            media = option.pop('media')
-            Option.objects.create(**option, media=media['name'], poll=poll)
-
         return poll