+++ polls/urls.py
@@ -1,6 +1,7 @@
 from rest_framework.routers import SimpleRouter
+from rest_framework_nested import routers
 
-from polls.views import PollViewSet
+from polls.views import PollViewSet, OptionViewSet
 
 
 app_name = 'polls'
@@ -9,6 +10,9 @@ router = SimpleRouter()
 
 router.register('', PollViewSet, basename='polls')
 
+option_router = routers.NestedSimpleRouter(router, '', lookup='poll')
+option_router.register('options', OptionViewSet, basename='poll-options')
+
 urlpatterns = [
 
-] + router.urls
+] + router.urls + option_router.urls