+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # CHANGELOG
 
+## v0.5.0 - 15.09.2020
+
+### Added
+
+* Option view set to make manipulations with options
+* Image to poll model
+* Description to poll model
+
+### Deleted
+
+* Logic to create poll with options
+* Options from poll serializer
+* Permission file, because now using build in from `rest_framework`
+
 ## v0.4.1 - 27.08.2020
 
 ### Changed